First commit
The application has 2 main views: the "index" which contains all the
application and the "login" view.

the Index view contains a HTML5/Angular webApp with 2 routes: home (the
homepage) and list (displays an ui.grid with data from the mongo
database).
the "list" route needs the user to be authentified.
the data displayed in the "list" route can be edited (CRUD operations)
only if the authentified user has the ADMIN role.